extracted wrapper exception
