+++ edelta.parent/edelta/src/edelta/interpreter/EdeltaInterpreter.java
@@ -106,24 +106,6 @@ public class EdeltaInterpreter extends XbaseInterpreter {
 		}
 	}
 
-	/**
-	 * Wraps a {@link Throwable} found in the result of interpretation
-	 */
-	public static class EdeltaInterpreterWrapperException extends RuntimeException {
-
-		private static final long serialVersionUID = 1L;
-		private final Exception exception;
-
-		public EdeltaInterpreterWrapperException(Exception exception) {
-			super(exception);
-			this.exception = exception;
-		}
-
-		public Exception getException() {
-			return exception;
-		}
-	}
-
 	public void setInterpreterTimeout(final int interpreterTimeout) {
 		this.interpreterTimeout = interpreterTimeout;
 	}
